Renamed some more dependencies
